Implement unified test wait operation

+++ src/libmongoc/src/mongoc/mongoc-crypt-private.h
@@ -48,7 +48,7 @@ _mongoc_crypt_new (const bson_t *kms_providers,
                    bool bypass_auto_encryption,
                    bool bypass_query_analysis,
                    mc_kms_credentials_callback creds_cb,
-                   int32_t cache_expiration_ms,
+                   int64_t cache_expiration_ms,
                    bson_error_t *error);
 
 void
